# lib/api-zod/src/generated/types/itemStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTT Stream Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemStats {
  uptime_pct: number;
  total_incidents: number;
  /** @nullable */
  mttr_seconds?: number | null;
  window: string;
}